requirements.txt
async-timeout==5.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.6.12
cryptography==44.0.2
idna==3.10
ifaddr==0.2.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
requests==2.32.3
urllib3==2.3.0
websockets==14.2
zeroconf==0.141.0
fastapi[standard]>=0.113.0,<0.116.0
pydantic>=2.7.0,<3.0.0
pybose
python-dotenv